+++ README.md
@@ -1,20 +1,21 @@
 # What is Maestro?
+
 Maestro is an open source file manager, designed to help users to have true control of their files.
 
 Maestro is a part of the Ship-In-30 project from [David's Developer Safeplace](https://discord.gg/devsafeplace) Discord server.
 
 ## Features
+
 Throughout the next 30 days, we as a team aim to add the following:
 
 - [ ] Basic file management (as a file manager would have :P)
-- [ ] An original yet interesting way to visualise storage space on a drive
+- [ ] An original yet interesting way to visualize storage space on a drive
 - [ ] A sleek, creative, and helpful user interface to make managing your storage much more efficient
 - [ ] A way to protect your valuable file contents with passwords and encryption
 
 ## Tech Stack
 
-Maestro uses React Native to interface managament of files, using Typescript for general functionality.
-
+Maestro uses React Native to interface management of files, using Typescript for general functionality.
 
 ## Contribution
 
@@ -34,14 +35,25 @@ Upon completing your changes **that are tested**, please raise a pull request th
 
 ### Tasks
 
-Within this project, contributors can help to work on tasks. Tasks will be split into different issues. To claim a task, find the issue that includes your chosen task and comment that you would like to work on a specific part of the issue. It's key that we communicate with each other to create a more productive and effecient development environment. You can also mention that you've claimed a task in the [discussion](https://discord.com/channels/368853404723707914/1073307477405335592) channel. 
+Within this project, contributors can help to work on tasks. Tasks will be split into different issues. To claim a task, find the issue that includes your chosen task and comment that you would like to work on a specific part of the issue. It's key that we communicate with each other to create a more productive and efficient development environment. You can also mention that you've claimed a task in the [discussion](https://discord.com/channels/368853404723707914/1073307477405335592) channel.
 
 Make sure to read all the discussion within the issue to know what needs to be completed, there will be a checklist of tasks and who's working on them in the issue description to help things flow smoothly.
 
-If you would like to unclaim a task, make sure to comment on the relevent issue to let everybody else know that the issue is open for contribution. 
+If you would like to unclaim a task, make sure to comment on the relevant issue to let everybody else know that the issue is open for contribution.
 
 You can also collaborate on a task with another contributor, just let the rest of the team know in the issue discussion.
 
 ---
 
 ![alt text](https://images-ext-1.discordapp.net/external/ZodgpNW25bKMHly3yapNdxjUcH8s__4xR5pqVxXy1dA/https/cdn-longterm.mee6.xyz/plugins/embeds/images/368853404723707914/668c5398d4a84cfbd3475ae6d201c456e398f819dcbf513c719b6d00aac67756.png?width=848&height=676)
+
+## How to run
+
+Clone the repo, cd into the project, and run `yarn install`
+
+### Mac Installation
+
+If you're on mac, you'll need to also install pods to get Xcode running properly. Do so with `npx pod-install`.
+
+After your dependencies are set up, go ahead and open the `ios/**.xcworkspace` file with Xcode.
+Xcode will now set up your project, after it's done loading, go ahead and click the play button on the Xcode UI, the app should automatically start in about 5 minutes. Go get a coffee!
