+++ navigation.d.ts
@@ -1,0 +1,7 @@
+import type {StackParamList} from './src/navigation/RootNavigation';
+
+declare global {
+  namespace ReactNavigation {
+    interface RootParamList extends StackParamList {}
+  }
+}
