+++ package.json
@@ -11,8 +11,12 @@
     "windows": "react-native run-windows"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
     "react": "18.2.0",
     "react-native": "0.71.0",
+    "react-native-safe-area-context": "^4.5.3",
+    "react-native-screens": "^3.21.0",
     "react-native-windows": "0.71.18"
   },
   "devDependencies": {
@@ -27,11 +31,11 @@
     "babel-jest": "^29.2.1",
     "eslint": "^8.19.0",
     "jest": "^29.2.1",
+    "metro-config": "^0.73.10",
     "metro-react-native-babel-preset": "0.73.5",
     "prettier": "^2.4.1",
     "react-test-renderer": "18.2.0",
-    "typescript": "4.8.4",
-    "metro-config": "^0.73.10"
+    "typescript": "4.8.4"
   },
   "jest": {
     "preset": "react-native"
