+++ src/components/Layout/Section.tsx
@@ -1,0 +1,55 @@
+import React from 'react';
+import {View, StyleSheet, Text} from 'react-native';
+import type {PropsWithChildren} from 'react';
+import {Colors} from 'react-native/Libraries/NewAppScreen';
+import {useColorScheme} from 'react-native';
+
+type SectionProps = PropsWithChildren<{
+  title: string;
+}>;
+
+// This is the default section component from React-Native, we could totally change it.
+export const Section = ({children, title}: SectionProps): JSX.Element => {
+  const isDarkMode = useColorScheme() === 'dark';
+  return (
+    <View style={styles.sectionContainer}>
+      <Text
+        style={[
+          styles.sectionTitle,
+          {
+            color: isDarkMode ? Colors.white : Colors.black,
+          },
+        ]}>
+        {title}
+      </Text>
+      <Text
+        style={[
+          styles.sectionDescription,
+          {
+            color: isDarkMode ? Colors.light : Colors.dark,
+          },
+        ]}>
+        {children}
+      </Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  sectionContainer: {
+    marginTop: 32,
+    paddingHorizontal: 24,
+  },
+  sectionTitle: {
+    fontSize: 24,
+    fontWeight: '600',
+  },
+  sectionDescription: {
+    marginTop: 8,
+    fontSize: 18,
+    fontWeight: '400',
+  },
+  highlight: {
+    fontWeight: '700',
+  },
+});
