+++ src/utils/constants.ts
@@ -1,0 +1,4 @@
+// Default object for screens in our app.
+export const DEFAULT_STACK_OPTIONS = {
+  headerShown: false, // Hiding the default React-Native header for a custom one we'll probably build.
+};
